split report detail page body
