+++ src/interfaces/business_monitor.ts
@@ -104,4 +104,32 @@ export const mockReports: IMockReport[] = [
     assuranceStandards: "確信準則 3000 號",
     isVerifiedByThirdParty: true,
   },
+  {
+    id: 6,
+    company: "環拓科技股份有限公司",
+    title: "2023 年永續報告書",
+    reportYear: "2023",
+    period: "2023/01/31 ~ 2023/12/31",
+    industry: "綠能環保",
+    capital: "無",
+    verificationAgency: "台灣檢驗科技股份有限公司(SGS)",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、TCFD、SASB)",
+    assuranceAgency: "無",
+    assuranceStandards: "無",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 7,
+    company: "環拓科技股份有限公司",
+    title: "2022 年永續報告書",
+    reportYear: "2022",
+    period: "2022/01/31 ~ 2022/12/31",
+    industry: "綠能環保",
+    capital: "無",
+    verificationAgency: "台灣檢驗科技股份有限公司(SGS)",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、TCFD、SASB)",
+    assuranceAgency: "無",
+    assuranceStandards: "無",
+    isVerifiedByThirdParty: true,
+  },
 ];
